change discountAmount type to Int

+++ Luncheon/SampleData/SampleFood.swift
@@ -16,12 +16,12 @@ struct SampleFood {
     let persianName:String
     let rate:Double
     let price:UInt
-    let discountAmount:Float
+    let discountAmount:Int
     let ingredients:[Ingredient]
     let image:UIImage
 
     
-    internal init(id: Int, persianName: String, rate: Double,price: UInt, discountAmount: Float, ingredients: [Ingredient], calorie: Int,image:UIImage) {
+    internal init(id: Int, persianName: String, rate: Double,price: UInt, discountAmount: Int, ingredients: [Ingredient], calorie: Int,image:UIImage) {
         self.id = id
         self.persianName = persianName
         self.rate = rate
